+++ src/serializer.hpp
@@ -10,9 +10,9 @@ struct Serializer {
 	static char *decodeAddress(char *mem, Address &e);
 	
 	struct Message {
-		static MessageHdr *allocEncodeJOINREQ(const Address &addr, int id, int ring_id, long heartbeat, size_t &msgSize);
-		static void decodeJOINREQ(MessageHdr *msg, Address &addr, int &id, int &ring_id, long &heartbeat);
-		// static MessageHdr *allocEncodeAD(const vector<MemberInfo> &lst, size_t &msgSize);
+		static MessageHdr *allocEncodeJOINREQ(const Address &addr, int id, int ring_id, int heartbeat, uint32_t &msgSize);
+		static void decodeJOINREQ(MessageHdr *msg, Address &addr, int &id, int &ring_id, int &heartbeat);
+		// static MessageHdr *allocEncodeAD(const vector<MemberInfo> &lst, uint32_t &msgSize);
 		// static void decodeAD(MessageHdr *msg, vector<MemberInfo> &lst);
 		static void dealloc(MessageHdr *msg) {
 			free(msg);
